src/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./requirements.in
#
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
cryptography==45.0.5
    # via pyopenssl
dnspython==2.7.0
    # via pymongo
docutils==0.21.2
    # via readme-renderer
id==1.5.0
    # via twine
idna==3.10
    # via requests
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.2.1
    # via keyring
keyring==25.6.0
    # via
    #   -r ./requirements.in
    #   twine
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
nh3==0.3.0
    # via readme-renderer
packaging==25.0
    # via twine
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via
    #   readme-renderer
    #   rich
pymongo==4.13.2
    # via -r ./requirements.in
pyopenssl==25.1.0
    # via -r ./requirements.in
python-dotenv==1.1.1
    # via -r ./requirements.in
readme-renderer==44.0
    # via twine
requests==2.32.4
    # via
    #   -r ./requirements.in
    #   id
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==14.0.0
    # via twine
tabulate==0.9.0
    # via -r ./requirements.in
twine==6.1.0
    # via -r ./requirements.in
typing-extensions==4.14.1
    # via pyopenssl
urllib3==2.5.0
    # via
    #   requests
    #   twine
wheel==0.45.1
    # via -r ./requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
